WA-450 Removed money; Added token amount
